# docs/swagger.yaml
basePath: /api/v1
definitions:
  core_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short-human readable message
        type: string
    type: object
...
  users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivanov Ivan
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+79998887766"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivanov Ivan
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 1
        type: integer
    type: object
  users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Ivanov Ivan
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 1
        type: integer
    type: object
  users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Ivanov Ivan
        type: string
      phone_number:
        example: "+79998887766"
        type: string
        x-nullable: true
    type: object
  users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Ivanov Ivan
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 1
        type: integer
    type: object
  users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivanov Ivan
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 1
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: kek
  title: Go rest prod API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Получить список пользователей с пагинацией
      parameters:
      - description: Лимит записей
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Получить список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: удалить пользователя в системе
      parameters:
      - description: ID удаляемого юзера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешное удаление
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Удалить пользователя
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Получить пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Найденный пользователь
          schema:
            $ref: '#/definitions/users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Получить пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующем в системе пользователе.
        ### Логика обновления полей (three-state logic):
        1. **Поле не передано**: `phone_number` игнорируется, значение в БД не меняется
        2. **Явно передано значение**: `"phone_number": "+711122233344"` — устанавливает новый номер
        3. **Передан null**: `"phone_number": null` — очищает поле в БД (set to NULL)
        Ограничения: `full_name` не может быть выставлен как null
      parameters:
      - description: ID изменяемого пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённый пользователь
          schema:
            $ref: '#/definitions/users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Изменение пользователя
      tags:
      - users
swagger: "2.0"